[frontend_python] Move code to src/frontend_python package
