+++ src/oif/frontend_python/__init__.py
@@ -1,0 +1,3 @@
+"""
+Subpackage for the user-facing Python code
+"""
